+++ bindings/react/src/components/RouterNavigator.jsx
@@ -185,6 +185,7 @@ class RouterNavigator extends BasicComponent {
       (route) => this.props.renderPage(route, this)
     );
 
+    node.swipeMax = this.props.swipePop;
     node.onDeviceBackButton = this.props.onDeviceBackButton || this._onDeviceBackButton.bind(this);
 
     this.update();
@@ -342,6 +343,30 @@ RouterNavigator.propTypes = {
    *  [ja][/ja]
    */
   animationOptions: PropTypes.object,
+
+  /**
+   * @name swipeable
+   * @type bool|string
+   * @required false
+   * @description
+   *  [en]
+   *  Enables swipe-to-pop functionality for iOS.
+   *  [/en]
+   *  [ja][/ja]
+   */
+  swipeable: PropTypes.oneOfType([PropTypes.bool, PropTypes.string]),
+
+  /**
+   * @name swipePop
+   * @type function
+   * @required false
+   * @description
+   *  [en]
+   *  Function called on swipe-to-pop. Must perform a popPage with the given options object.
+   *  [/en]
+   *  [ja][/ja]
+   */
+  swipePop: PropTypes.func,
 
   /**
    * @name onDeviceBackButton