+++ core/src/elements/ons-range.js
@@ -15,7 +15,8 @@ import BaseInputElement from './base/base-input';
 
 const scheme = {
   '': 'range--*',
-  '.range__input': 'range--*__input'
+  '.range__input': 'range--*__input',
+  '.range__focus-ring': 'range--*__focus-ring'
 };
 
 const activeClassToken = 'range__input--active';
@@ -49,6 +50,10 @@ export default class RangeElement extends BaseInputElement {
   constructor() {
     super();
 
+    this._boundOnMouseDown = this._onMouseDown.bind(this);
+    this._boundOnMouseUp = this._onMouseUp.bind(this);
+    this._boundOnTouchStart = this._onTouchStart.bind(this);
+    this._boundOnTouchEnd = this._onTouchEnd.bind(this);
     this._boundOnInput = this._update.bind(this);
     this._boundOnDragstart = this._onDragstart.bind(this);
     this._boundOnDragend = this._onDragend.bind(this);
@@ -61,8 +66,22 @@ export default class RangeElement extends BaseInputElement {
 
   /* Inherited props */
 
+  get _focusRing() {
+    return this.children[1];
+  }
+
   _update() {
-    this._input.style.backgroundSize = (100 * this._ratio) + '% 2px';
+    const input = this._input;
+
+    input.style.backgroundSize = `${100 * this._ratio}% 2px`;
+    this._focusRing.value = this.value;
+
+    // NOTE: "_zero" attribute is used for CSS styling.
+    if ((input.min === '' && input.value === '0') || input.min === input.value) {
+      input.setAttribute('_zero', '');
+    } else {
+      input.removeAttribute('_zero');
+    }
   }
 
   get _scheme() {
@@ -72,6 +91,7 @@ export default class RangeElement extends BaseInputElement {
   get _template() {
     return `
       <input type="${this.type}" class="${this._defaultElementClass}__input">
+      <input type="range" class="range__focus-ring" tabIndex="-1">
     `;
   }
 
@@ -85,14 +105,37 @@ export default class RangeElement extends BaseInputElement {
 
   /* Own props */
 
+  _onMouseDown(e) {
+    this._input.classList.add(activeClassToken);
+    setImmediate(() => this._input.focus());
+  }
+
+  _onTouchStart(e) {
+    this._onMouseDown();
+  }
+
+  _onMouseUp(e) {
+    this._input.classList.remove(activeClassToken);
+  }
+
+  _onTouchEnd(e) {
+    this._onMouseUp(e);
+  }
+
   _onDragstart(e) {
-    e.stopPropagation();
+    e.consumed = true;
     e.gesture.stopPropagation();
     this._input.classList.add(activeClassToken);
+    this.addEventListener('drag', this._onDrag);
+  }
+
+  _onDrag(e) {
+    e.stopPropagation();
   }
 
   _onDragend(e) {
     this._input.classList.remove(activeClassToken);
+    this.removeEventListener('drag', this._onDrag);
   }
 
   get _ratio() {
@@ -126,12 +169,20 @@ export default class RangeElement extends BaseInputElement {
   }
 
   connectedCallback() {
+    this.addEventListener('mousedown', this._boundOnMouseDown);
+    this.addEventListener('mouseup', this._boundOnMouseUp);
+    this.addEventListener('touchstart', this._boundOnTouchStart);
+    this.addEventListener('touchend', this._boundOnTouchEnd);
     this.addEventListener('dragstart', this._boundOnDragstart);
     this.addEventListener('dragend', this._boundOnDragend);
     this.addEventListener('input', this._boundOnInput);
   }
 
   disconnectedCallback() {
+    this.removeEventListener('mousedown', this._boundOnMouseDown);
+    this.removeEventListener('mouseup', this._boundOnMouseUp);
+    this.removeEventListener('touchstart', this._boundOnTouchStart);
+    this.removeEventListener('touchend', this._boundOnTouchEnd);
     this.removeEventListener('dragstart', this._boundOnDragstart);
     this.removeEventListener('dragend', this._boundOnDragend);
     this.removeEventListener('input', this._boundOnInput);