+++ core/src/elements/ons-switch.js
@@ -137,7 +137,7 @@ export default class SwitchElement extends BaseCheckboxElement {
       return;
     }
 
-    e.stopPropagation();
+    e.consumed = true;
 
     ModifierUtil.addModifier(this, 'active');
     this._startX = this._locations[this.checked ? 1 : 0];// - e.gesture.deltaX;
@@ -147,7 +147,8 @@ export default class SwitchElement extends BaseCheckboxElement {
   }
 
   _onDrag(e) {
-    e.gesture.srcEvent.preventDefault();
+    e.stopPropagation();
+    e.gesture.preventDefault();
     this._handle.style.left = this._getPosition(e) + 'px';
   }