+++ CHANGELOG.md
@@ -5,28 +5,39 @@ CHANGELOG
 dev
 ----
 
+### New Features
+
+ * core: Added `ons.modifier` object with a new set of methods to alter `modifier` attributes.
+ * ons-modal: Added `lift` animation ([#2078](https://github.com/OnsenUI/OnsenUI/pull/2078)).
+ * ons-navigator: iOS Swipe-to-pop.
+
 ### Bug Fixes
 
  * css-components, ons-list-item: Fixed [#2100](https://github.com/OnsenUI/OnsenUI/issues/2100).
+ * ons-list: Fixed double top border when located right under toolbar.
+ * Swipeable components: Swipe is now smoother.
+ * Swipeable components: Fixed compatibility of some swipeable components.
 
-### New Features
+### Misc
 
  * core: Added `ons.modifier` object with a new set of methods to alter `modifier` attributes.
  * ons-modal: Added `lift` animation ([#2078](https://github.com/OnsenUI/OnsenUI/pull/2078)).
  * Added ripple effects properly on `ons-range`, `ons-checkbox`, `ons-radio-button`, `ons-toolbar-button` and `ons-alert-dialog` buttons for Material Design.
  * Added `ons-alert-dialog-button` elements.
  * Added `size` attribute on `ons-ripple` elements.
+ * css-components: Adjusted font size of `.back-button__label` (improves iOS slide animation).
 
 v2.4.2
 ----
 
-### Bug Fixes
-
- * Revert a recent change that creates `ons-tabbar` issues. Fixed [#2082](https://github.com/OnsenUI/OnsenUI/issues/2082).
-
 ### New Features
 
  * ons-fab: Added new appearances for iOS.
+ * ons-ripple: Added `size` attribute.
+
+### Bug Fixes
+
+ * Revert a recent change that creates `ons-tabbar` issues. Fixed [#2082](https://github.com/OnsenUI/OnsenUI/issues/2082).
 
 v2.4.1
 ----